Move types in own files
